+++ backend_api/app/applications/users/crud.py
@@ -1,1 +1,18 @@
-# 232
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy import select
+
+from applications.auth.password_handler import PasswordEncrypt
+from applications.users.models import User
+
+async def create_user_in_db(email, name, password, session: AsyncSession):
+    hashed_password = await PasswordEncrypt.get_password_hash(password)
+    new_user = User(email=email, name=name, hashed_password=hashed_password)
+    session.add(new_user)
+    await session.commit()
+
+
+
+async def get_user_by_email(email, session: AsyncSession) -> User | None:
+    query = select(User).filter(User.email == email)
+    result = await session.execute(query)
+    return result.scalar_one_or_none()
