+++ backend_api/app/applications/users/router.py
@@ -1,20 +1,11 @@
 from fastapi import APIRouter, Depends, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from applications.auth.password_handler import PasswordEncrypt
-from applications.users.models import User
+from applications.users.crud import create_user_in_db
 from applications.users.schemas import BaseFields, RegisterUserFields
 from database.session_dependencies import get_async_session
 
 router_users = APIRouter()
-
-
-async def create_user_in_db(email, name, password, session: AsyncSession):
-    hashed_password = await PasswordEncrypt.get_password_hash(password)
-    new_user = User(email=email, name=name, hashed_password=hashed_password)
-    session.add(new_user)
-    await session.commit()
-
 
 @router_users.post("/create", status_code=status.HTTP_201_CREATED)
 async def create_user(new_user: RegisterUserFields, session: AsyncSession = Depends(get_async_session)) -> BaseFields:
